pyproject.toml
[project]
name = "native-audio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.2",
    "numpy>=2.3.5",
    "openai>=2.13.0",
    "python-dotenv>=1.2.1",
    "soundfile>=0.13.1",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[dependency-groups]
dev = [
    "ruff>=0.14.9",
]
